+++ resources/views/admin/asset/index.php
@@ -3,27 +3,26 @@
 declare(strict_types=1);
 
 /**
- * @see AbstractAssetController::actionIndex()
+ * @see AssetController::actionIndex()
  *
  * @var View $this
- * @var AssetModelInterface $model
- * @var AssetArrayDataProvider $provider
+ * @var File $file
  */
 
-use Hirtz\Media\Models\Interfaces\AssetModelInterface;
-use Hirtz\Media\Modules\Admin\Controllers\AbstractAssetController;
-use Hirtz\Media\Modules\Admin\Data\AssetArrayDataProvider;
-use Hirtz\Media\Modules\Admin\Widgets\Grids\AssetGridView;
-use Hirtz\Media\Modules\Admin\Widgets\Navs\AssetModelActionDropdown;
-use Hirtz\Media\Modules\Admin\Widgets\Navs\AssetModelHeader;
+use Hirtz\Media\Models\File;
+use Hirtz\Media\Modules\Admin\Controllers\AssetController;
+use Hirtz\Media\Modules\Admin\Widgets\Grids\FileAssetGridView;
+use Hirtz\Media\Modules\Admin\Widgets\Navs\FileHeader;
+use Hirtz\Media\Modules\Admin\Widgets\Navs\FileSubmenu;
 use Hirtz\Skeleton\Web\View;
 use Hirtz\Skeleton\Widgets\Grids\GridContainer;
 
-echo AssetModelHeader::make()
-    ->model($model)
-    ->content(AssetModelActionDropdown::make()
-        ->provider($provider));
+echo FileHeader::make()
+    ->model($file);
+
+echo FileSubmenu::make()
+    ->model($file);
 
 echo GridContainer::make()
-    ->grid(AssetGridView::make()
-        ->provider($provider));
+    ->grid(FileAssetGridView::make()
+        ->file($file));