+++ src/Modules/Admin/Widgets/Grids/FileAssetGridView.php
@@ -128,11 +128,6 @@ class FileAssetGridView extends GridView
      */
     protected function can(string $action, Asset $asset): bool
     {
-        $model = $asset->model;
-
-        return $this->webuser->can($asset->getPermissionName($action), [
-            'asset' => $asset,
-            $model->getParamName() => $model,
-        ]);
+        return $this->webuser->can($asset->getPermissionName($action), ['asset' => $asset]);
     }
 }