+++ src/Modules/Admin/Widgets/Navs/AssetActionDropdown.php
@@ -76,11 +76,6 @@ class AssetActionDropdown extends ActionDropdown
 
     protected function canDeleteAsset(): bool
     {
-        $model = $this->model->model;
-
-        return $this->webuser->can($this->model->getPermissionName('delete'), [
-            'asset' => $this->model,
-            $model->getParamName() => $model,
-        ]);
+        return $this->webuser->can($this->model->getPermissionName('delete'), ['asset' => $this->model]);
     }
 }